Avance Contactos, Login, Sistema de Roles, Seeder de Prueba y Vistas
Se trabajó en las vista principales de Login, con sus logos y además se crearon las vista necesarias para manejo del CRUD de Contactos. Por ultimo se resuelven errores del Sistema de Roles, además se crean Seeders de datos de prueba

+++ database/migrations/2025_06_02_023524_create_organizaciones_table.php
@@ -22,7 +22,7 @@ return new class extends Migration
                 $table->string('urlPageWeb', 200)->nullable();
                 $table->unsignedBigInteger('provincia_id');
                 $table->foreign('provincia_id')->references('id')->on('provincias');
-                $table->unsignedBigInteger('canton_id');
+                $table->unsignedSmallInteger('canton_id');
                 $table->foreign('canton_id')->references('id')->on('cantones');
                 $table->unsignedBigInteger('distrito_id');
                 $table->foreign('distrito_id')->references('id')->on('distritos');
